Move channel to main i2s.rs.
This moves the channel implementation (which turns out to be
generic) to the main i2s.rs file in preparation for PDM and TDM mode
implementation.
